refactor(commons): lift _semver and the artifact reconstructors
A pure move, deferred from the Task 3 review. The version-floor parser both
Neo4j probes compare against their _ANALYZER_FLOOR is commons.backend.semver;
the three reconstructors of the shared artifact layer (artifact / config_key /
dependency -- the one part of the graph every analyzer projects identically,
into the same Py* models the generic ABC promises) are commons/artifacts.py.
The Python backend and its reconstruct module re-import them under the old
names, pinned by identity in tests/analysis/commons/test_lifted_helpers.py;
the TypeScript Neo4j backend imports from commons instead of from the Python
package. The TS backend's private _module(props, children) helper is renamed
_build_module so a grep for the retired _module node property finds prose only.
